test(client): cover the menu recipes on rendered Joy components
The suite read keys off returned objects, so the module's strongest
claim - that our selected-row declaration outranks the rule Joy paints
the same row from - had no test. Renders a Joy Select and a Joy Menu
and asserts computed backgrounds instead.

jsdom runs the real cascade over Emotion's stylesheet, so a rule losing
its specificity race changes a computed value. It does not substitute
var(), which is what makes "Joy won" legible: the value reads back as
the unresolved var(--variant-plainActiveBg, ...) string.

Replaces the inert-on-a-Menu proxy. It filtered returned keys for
`role=` and asserted one entry, which would have passed if the recipe
grew a '& .MuiMenuItem-root' key. Injecting exactly that key proved the
first render test missed it too: Joy's own .Mui-selected rule wins the
source-order tie on the SELECTED row, so the leak only shows on the
plain rows. Both rows are asserted, and a contrast case renders the
same menu with menuItemListSx to show the assertion can fail.

Also covers the corner constant and the threaded radius, and says what
the MENU_INSET equality test guards - both sides read one constant, so
it cannot fail while they do; it guards against either site reverting
to a literal.

menuRowSx coverage stays as it landed with the press-tint fix, which
already exercises the default and danger arms; this adds its corner.
